Tercer commit: Modifico el metodo sumar para que sea con double

+++ GestionGit/src/gestionGit/Operaciones.java
@@ -2,11 +2,13 @@ package gestionGit;
 
 public class Operaciones {
 
-	public int sumar(int a, int b) {
+	public double sumar(int a, int b) {
 		return a + b;
 	}
 	
 	public int restar (int a,int b) {
 		return a-b;
 	}
+	
+	
 }
